--- server/src/modules/module-0/exercise/solution.js
@@ -1,7 +1,0 @@
-function greetUser(name) {
-    return `Hello, ${name}! Welcome to Node.js!`;
-}
-
-module.exports = {
-    greetUser
-};
