--- server/src/modules/module-0/module.json
@@ -1,31 +1,0 @@
-{
-  "id": "module-0",
-  "title": "Environment Setup & Node.js Basics",
-  "description": "Learn the fundamentals of Node.js, including runtime environment, modules, package management, and file system operations",
-  "difficulty": "Beginner",
-  "estimatedTime": "45 minutes",
-  "tags": ["nodejs", "basics", "modules", "npm", "filesystem"],
-  "files": {
-    "lab": {
-      "readme": "lab/README.md"
-    },
-    "exercise": {
-      "readme": "exercise/README.md",
-      "server": "exercise/greetings.js",
-      "test": "exercise/test.js",
-      "solution": "exercise/solution.js",
-      "package": "exercise/package.json"
-    }
-  },
-  "learningObjectives": [
-    "Understand Node.js runtime environment",
-    "Learn module system and exports",
-    "Master package management with NPM",
-    "Practice file system operations",
-    "Create and use custom modules"
-  ],
-  "prerequisites": [
-    "Basic JavaScript knowledge",
-    "Understanding of functions and variables"
-  ]
-}
