+++ server/src/modules/module-2/exercise/server.js
@@ -1,0 +1,10 @@
+const http = require('http');
+
+// TODO: Implement your server here
+// 1. Create HTTP server
+// 2. Handle GET requests to "/"
+// 3. Return JSON response
+// 4. Handle 404 for other routes
+// 5. Listen on port 3000
+
+console.log('Server starting...');
